docs: Update package-lock.json.
This is the result of just running `npm install` in this Hexo docs
portion of the Apollo iOS repository.  It is the only docs repository
that is demonstrating this out-of-sync `package-lock.json` state, so I'm
not too worried about this commit.  Generally speaking though, this
shouldn't have happened on an `npm install`.