Be defensive about the scope of Renovate.
By default Renovate will scan the repository for all `package.json` files and update those dependencies.

This might be great for some (unlikely) future Node package within this repository, but the current goal is to enable these updates for the docs themselves and nothing else.  In order to avoid Renovate operating unexpectedly on packages outside the `docs` scope, this adds the Renovate `packageFiles` declaration directly to the root `renovate.json` with the sole desired `package.json`: `docs/package.json`.