Include `PLATFORM_NAME` in bundle identifier
As of ~5 days ago, App Store submissions will fail with the ITMS-90806 publishing error code if there are multiple bundles with the same identifier. This is a problem for apps that e.g. include an iOS version alongside a watchOS version.

This change just appends the `PLATFORM_NAME` to the bundle identifier to get around the issue.

Here's a thread from Apple's message boards about the issue: https://forums.developer.apple.com/thread/122048