Store date `Record` was last received in SQLite table
Motivation:

Any time a developer wants to cache data they inevitably also want a way to invalidate the cache or know when to ignore its data.

In order to support those use cases, the date that the data was last verified as received needs to be stored.

Modifications:

- Add: `RecordRow` struct that becomes the underlying storage base type for `RecordSet` that contains the `Record` and `lastReceivedAt` date
- Add: Mechanism to `SQLIteNormalizedCache` to handle schema migrations

Result:

TTL and cache age will be soon possible to query based on the data's `lastReceivedAt` date.