Refactor TestCacheProvider to enable more flexible initialization
This exposes a synchronous makeNormalizedCache() method on cache dependent test cases, which replaces the existing withCache() block. That allows us to use it from setUp/tearDown and clean up the test structure.

Right now only the watcher tests use this, but I  plan on migrating the other cache dependent tests as well.