Add context object to `GraphQLResult`
Motivation:

`RecordRow.lastReceivedAt` is an important property that developers want access to in order to inspect
the age of a given result.

In order to provide this, it needs to be returned in GraphQL results in some fashion.

Modifications:

- Add: `GraphQLResultContext` object for holding result metadata, such as the `resultAge`
- Add: `context` property to `GraphQLResult` that allows access to the result metadata
- Change: `NormalizedCache.loadRecords` completion to receive `RecordRow?` instead of `Record?`
- Change: `RecordSet` to use `RecordRow` rather than `Record`
- Change: `SQLiteNormalizedCache.init(fileURL:shouldVacuumOnClear:)` to be a convenience initializer
- Change: `SQLIteNormalizedCache` initializers to accept an optional `RecordSet`

Result:

Developers can now access a `context` property on `GraphQLResult` to inspect metadata behind a given result.