Add cache fallback CachePolicy
Motivation:

It is sometimes acceptable to fallback to whatever cache data might be available when a particular fetch request fails either due to network or server problems.

This adds a missing 'CachePolicy' that falls back to use the cache after first attempting a fetch request.

Modifications:

- Add: `fetchReturningCacheDataOnError` case to `CachePolicy`
- Add: `CacheError` type to `LegacyCacheReadInterceptor` that is returned if both the network and cache fails to provide a successful result
- Change: `LegacyCacheReadInterceptor` to handle new policy

Result:

Developers should now have the ability to rely on their cache in more scenarios, such as when the server or network is having issues.