Use simplified IR for now
This switches to a simplified IR that more directly corresponds to the structure of the GraphQL input.

I'm working on an elaboration of this that is more suitable to generate code from because it encodes type conditions and @skip/@include in a way that allows for efficient generation of the corresponding response shapes (collecting and merging fields where appropriate). That'll have to come in a separate PR.