Define all properties on JavaScript objects as read-only
We currently use `lazy` to initialize properties on demand, but that means changes to those stored property values won't be reflected in the wrapped JavaScript object. To avoid misunderstanding, I've made all projects `private(set)` for now.

If we wanted full two-way mapping between these wrappers and the underlying JavaScript objects, we'd have to define computed properties with both getters and setters. That's a lot of extra code for something we don't need right now, and if we do it's probably better solved through a property wrapper.