Invert Dependency on JavaScriptBridge
JavaScriptObjects now retain JavaScriptBridge
JavaScriptBridge `wrapperMap` now has wear references to `JavaScriptObject`s

Rename `ApolloCodgenFrontend` to `GraphQLJSFrontend`

I kept the JavaScript frontend as `ApolloCodgenFrontend` because it's not just `graphql-js`. It's a wrapper for graphql-js that specifically exposes the necessary pieces for the Apollo Codegen and has a few functions that handle parsing the graphls-js return results into the compilation result.