Enable ApolloPerformanceTests
Closes #2387

The issue mentioned that re-enabling the tests was waiting on #2386 (Update internal testing APIs to use 1.0 code generation).
Since that was already completed I uncommented the tests to see if they would run, and they run with no issues.

Only ran into one issue in which I had to update `query.operationName` to `Query.operationName` due to the following error:
Static member 'operationName' cannot be used on instance of type 'Query'