Update docs to remove the leading period from the `.embeddedInTarget` command line option
I initially ran the `apollo-ios-cli init` command with `--module-type .embeddedInTarget` but got an error indicating that I shouldn't include the period:

```
Error: The value '.embeddedInTarget' is invalid for '--module-type <module-type>'. Please provide one of 'embeddedInTarget', 'swiftPackageManager' or 'other'.
```

(I follow that this is because the enum member that it maps to is `ModuleType.embeddedInTarget` but the user doesn't really know that yet.)