Change interceptors back to being created all at once
I had changed the interceptors to being initialized on demand so that we could prevent initializing interceptors that ended up not being needed. But this meant that we were re-initializing the interceptors on a retry. That prevented them from carrying state over to the retry, which was not intended. It broke the MaxRetryInterceptor completely.

This could have been fixed by having the MaxRetryInterceptor use TaskLocal values for state, but that’s not the API we want to provide publicly. User created interceptors should be able to hold state.

So I moved back to providing all interceptors during initialization of the `RequestChain`.

I had also changed the interceptor provider to passing through the `GraphQLRequest` instead of just the `GraphQLOperation`. When we were getting providers on demand, that meant that we could pass through the request as it had been mutated by previous parts of the request chain. But since we are no longer doing that, we shouldn’t be providing users the request since it will only be the inital request and that could be unclear.

Instead, we should go back to the 1.0 behavior or just providing the immutable `GraphQLOperation.